Heart desease prediction
Over the last decade heart disease is the main reason for death in the world. Almost one person dies of Heart disease about every minute in India alone. In order to lower the number of deaths from heart diseases, there has to be a fast and efficient detection technique. Decision Tree is one of the effective data mining methods till this date. The algorithm used in this project is namely are Decision Tree, k-nearest neighbours algorithm (KNN).Heart disease defines several healthcare conditions that are vast in nature which is related to the heart and has many basic causes that affect the entire body. The existing data of heart disease patients from heart dataset taken from kaggle is used to make a test and clearance to the performance of decision tree and KNN algorithms.These datasets consist of 303 instances and 14 attributes.This study's goal is to predict the heart disease in single patient and multipatients.In this project, of machine learning algorithm for predicting the heart diseases of a person. The dataset for this project is manually take the values for single patients and automatically taken from the dataset values for multipatients. With this project, we have shown that the raw data from the heart dataset actually increases the accuracy of the model using decision tree and KNN machine learning algorithm.